Begin abstracting the streaming API functions
