+++ package.json
@@ -24,6 +24,7 @@
     "react-dom": "^18.2.0",
     "react-markdown": "^8.0.6",
     "remark-gfm": "^3.0.1",
+    "uuid": "^9.0.0",
     "zustand": "^4.3.7"
   },
   "devDependencies": {
@@ -31,6 +32,7 @@
     "@types/prismjs": "^1.26.0",
     "@types/react": "^18.0.33",
     "@types/react-dom": "^18.0.11",
+    "@types/uuid": "^9.0.1",
     "eslint": "^8.37.0",
     "eslint-config-next": "^13.3.0",
     "prettier": "^2.8.7",
